serverRPI/requirements.txt
Flask==1.1.2
Flask-Cors==3.0.10
Flask-JWT==0.3.2
Flask-Login==0.5.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.5.1
future==0.18.2
PyJWT==1.4.2
SQLAlchemy==1.4.11
requests==2.25.1